+++ src/index.js
@@ -1,6 +1,7 @@
 import './style.css';
 import { addBtn, input } from './modules/Constants.js';
 import { renderTasks, saveAndRender } from './modules/renderTasks.js';
+import { toggleDisplay, createTask } from './modules/utils.js';
 
 let tasksArr = [];
 if (localStorage.length > 0) {
@@ -15,32 +16,14 @@ window.removeTask = (_id) => {
   saveAndRender(tasksArr);
 };
 
-const createTask = (task) => {
-  const index = tasksArr.length;
-  return {
-    description: task,
-    completed: false,
-    id: index,
-  };
-};
-
 addBtn.addEventListener('click', () => {
   if (input.value !== '') {
-    const taskObj = createTask(input.value);
+    const taskObj = createTask(input.value, tasksArr.length);
     tasksArr.push(taskObj);
     saveAndRender(tasksArr);
     input.value = '';
   }
 });
-
-const toggleDisplay = (elemGroup1, elemGroup2) => {
-  elemGroup1.forEach((elem) => {
-    elem.classList.toggle('d-none');
-  });
-  elemGroup2.forEach((elem) => {
-    elem.classList.toggle('d-none');
-  });
-};
 
 window.editTask = (id) => {
   const normalDisplay = document.querySelector(`[data-id="${id}"]`).querySelectorAll('.normal-display');
